pyproject.toml
[project]
name = "inadyn-ddns-aws"
version = "0.1.0"
description = "AWS CDK stack for In-a-Dyn compatible Route 53 DDNS"
authors = ["Petras Petroskevicius <p.petroskevicius@gmail.com>"]
dependencies = ["aws-cdk-lib==2.177.0", "constructs>=10.0.0,<11.0.0", "boto3"]
requires-python = ">=3.9"
